requirements.txt
discord.py-self
aiohttp
websockets
colorama
requests
psutil
disnake
spotipy